pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "interactive-utils"
version = "0.1.0"
description = "Utilities for defining bounding boxes and related interactive geospatial tasks."
authors = [
    { name="Your Name", email="your.email@example.com" }
]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "numpy",
    "matplotlib",
    "scikit-learn",
    "ipyleaflet",
    "ipywidgets",
    "ipykernel",
    "rasterio",
    "geotessera>=0.7.0",
    "pyyaml",
    "tqdm",
]

[project.urls]
Homepage = "https://github.com/yourusername/interactive-utils"
Repository = "https://github.com/yourusername/interactive-utils"
Issues = "https://github.com/yourusername/interactive-utils/issues"

[tool.setuptools]
packages = ["interactive"]

[tool.setuptools.package-data]
"interactive" = ["py.typed"]

[tool.setuptools.exclude-package-data]
"interactive" = ["*.pyc", "__pycache__"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"

[tool.mypy]
python_version = 3.8
ignore_missing_imports = true
strict = true

[tool.uv.sources]
geotessera = { path = "../geotessera" }
